% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igebase.R
\name{igebase}
\alias{igebase}
\title{Download data from the IGE API}
\usage{
igebase(igebase_ID, code_region = NA, region = NA, POSIXct = FALSE,
  label = FALSE, startdate, enddate, freq, show_metadata = FALSE,
  cache)
}
\arguments{
\item{igebase_ID}{A character string or numeric value. The \code{ID} code of the requested table.
This \code{ID} corresponds to the \code{ID} column from \code{\link{cache}}.}

\item{code_region}{Character string or regular expression to be matched. It is used to find a region using
the province or municipality code is provided by INE. This parameter can not be use with \code{region}.}

\item{region}{Character string or regular expression to be matched. It is used to find a region by its name.
This parameter can not be used with \code{code_region}.}

\item{POSIXct}{if \code{TRUE}, two new columns, \code{periodo} and \code{periodicidade} are added.
Useful for \code{freq} filter, if \code{FALSE}, these fields are not added.}

\item{label}{if \code{TRUE}, the label of te requested dataframe is returned, if \code{FALSE} it is not returned.}

\item{startdate}{Character string. Must be in "yyyy-mm-dd" form.}

\item{enddate}{Character string. Must be in "yyyy-mm-dd" form.}

\item{freq}{Character string. For fetching yearly ("anual"), quaterly ("trimestral"), monthly ("mensual") values.}

\item{show_metadata}{if \code{FALSE}, only get the data frame
if \code{TRUE}, show information of metadata, used in \code{\link{igebase_get}}.}

\item{cache}{Data frame with tables from IGE API.}
}
\value{
Data frame with all available requested data.
}
\description{
This function downloads the requested information using the IGE API.
}
\examples{
# Informaci贸n climatol贸xica por meses e estaci贸ns meteorol贸xicas.
df <- igebase(1651, POSIXct = TRUE, label = TRUE, freq = "anual", show_metadata = TRUE)

# Autosuficiencia en emprego using 'code_region'.
df <- igebase(7405, code_region ="15|12", label = TRUE)

# Autosuficiencia en emprego using 'region'.
df <- igebase(7405, region ="Galicia|Vigo")

}
