% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igebase_get.R
\name{igebase_get}
\alias{igebase_get}
\title{Call the IGE API and return a formatted data frame}
\usage{
igebase_get(igebase_ID, show_metadata = FALSE, cache)
}
\arguments{
\item{igebase_ID}{A character string. The \code{ID} code of the requested table.}

\item{show_metadata}{if \code{FALSE}, only get the data frame
if \code{TRUE}, show information of metadata.
Normally gived by the \code{\link{igebase_search}} function.}

\item{cache}{Data frame with tables from IGE API.}
}
\value{
A data frame.
}
\description{
This function calls the IGE API, capture the data in CSV
format and it is formatted as data frame. To be used inside of \code{\link{igebase}}
}
\examples{
# Temperaturas e insolación.
df.temp <- igebase_get(520)
df.temp <- igebase_get(520, show_metadata = TRUE) # see information of metadata


}
