% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igebase_search.R
\name{igebase_search}
\alias{igebase_search}
\title{Search information about data available through IGE API}
\usage{
igebase_search(pattern, fields = "label", extra = FALSE,
  exact = FALSE, cache)
}
\arguments{
\item{pattern}{Character string or regular expression to be matched.}

\item{fields}{Character vector of column name through which to search.}

\item{extra}{if \code{FALSE}, only the ID and label are returned,
if \code{TRUE}, all columns of the \code{cache} are returned.}

\item{exact}{Search for the exact pattern.}

\item{cache}{Data frame with metadata about API and IGE information.}
}
\value{
Data frame with metadata that match the search term.
}
\description{
This funcion allows finds those tables that match the search term and
returns a data frame with results
}
\examples{
igebase_search("temperatura")
igebase_search("temperatura", extra = TRUE)

# with regular expression operators
igebase_search(pattern = "temperatura|poboación")

# with exact expression
igebase_search("Temperaturas e insolación", exact = TRUE)


}
