% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igebaseperiodos2POSIXct.R
\name{igebasecodtempo2POSIXct}
\alias{igebasecodtempo2POSIXct}
\title{Add a POSIXct dates to a IGE base API return}
\usage{
igebasecodtempo2POSIXct(df)
}
\arguments{
\item{df}{data frame returned from API call}
}
\value{
The original data frame with two new columns, \code{periodo} and \code{periodicidade} is returned.
}
\description{
Add a POSIXct date column as well as a column with the
appropiate granularity to a IGE base API return
}
